Feature/#152: SystemMediumWidgetView UI 구현
